Refactor: include release map in module exports
