+++ package.json
@@ -6,7 +6,7 @@
   "license": "GPL-3.0-only",
   "scripts": {
     "clean": "rimraf ./lib && rimraf ./lib-esm",
-    "build": "npm run compile && npm run compile-esm && npm run copy-declarations && npm run build-docs",
+    "build": "npm run clean && npm run compile && npm run compile-esm && npm run copy-declarations && npm run build-docs",
     "test": "npm run lint && npm run typecheck && jest",
     "lint": "eslint src/**/*.ts scripts/**/*.ts",
     "typecheck": "tsc --noEmit",
