+++ src/constants.ts
@@ -16,6 +16,16 @@ export enum ColonyVersion {
   CeruleanLightweightSpaceship = 6,
 }
 
+// These are the corresponding git release tags for the deployed versions of the Colony Network
+export const releaseMap = {
+  [ColonyVersion.GoerliGlider]: 'glider-rc.1',
+  [ColonyVersion.Glider]: 'glider',
+  [ColonyVersion.AuburnGlider]: 'auburn-glider',
+  [ColonyVersion.BurgundyGlider]: 'burgundy-glider',
+  [ColonyVersion.LightweightSpaceship]: 'lwss',
+  [ColonyVersion.CeruleanLightweightSpaceship]: `5acb79d6e441a2ca14330b9419cbfdaa9c5b56f3`,
+};
+
 /**
  * The newest colony version. This will be used when deploying new colonies */
 export const CurrentVersion = ColonyVersion.CeruleanLightweightSpaceship;
